Clean up `getScore` by leveraging the new switch expression and extracting a helper method (the for loop was effectively just running twice, once for each player). Reduced scope of variables and made each branch of the `if` return a score. Rename score fields to align with player name fields.
I suspect this is what the kata was going for as far as a cleaned up result. Tbh, it's not bad at all, and probably a good enough point to stop. The code is clear enough, and unless the rules of tennis change, there's not really any need to fiddle with those switches and conditions.